Use version 1.22 of the dynamo-local container
Version 2 doesn't seem to like authenticating. I swear it worked
initially, but that image was updated before I even made my first commit
and on a re-run today, all the tests fail because it can't auth.

It's not important to demonstrating the point here, so I'm changing the
version and moving on.